fix migrations on fresh database
the migration did only work on existing databases where the
ContentType table was already exising while executing this
migrations. Now it is working with and without existing
database.

For installations that have executed only the old migration 0002
the migration 0003 does nothing and the migration 0004 just
alter the fields to the definition they already are using
so nothing happens